refactor and move lock tasks
