+++ backend/infrahub/workflows/catalogue.py
@@ -537,7 +537,7 @@ CLEAN_UP_DEADLOCKS = WorkflowDefinition(
     name="clean_up_deadlocks",
     type=WorkflowType.INTERNAL,
     cron="* * * * *",
-    module="infrahub.pools.tasks",
+    module="infrahub.locks.tasks",
     function="clean_up_deadlocks",
     concurrency_limit=1,
     concurrency_limit_strategy=ConcurrencyLimitStrategy.CANCEL_NEW,
