+++ changelog/1290.added.md
@@ -1,0 +1,1 @@
+Clean up deadlocks at set intervals controlled by the `INFRAHUB_CACHE_CLEAN_UP_DEADLOCKS_INTERVAL_MINS` environment variable with a default value of `15` mins
